Streamlining moving account handling to a base class b/c it does not belong to toolbar or drawer implementation
